test: amend wall test scenario

+++ game-runner/main.cpp
@@ -90,7 +90,7 @@ void wallTest() {
     Wall *snake2[64];
     Wall *snake3[64];
     for (int x = 0; x < 16; x++) {
-        walls[x] = new Wall(x * 4, 0, 4);
+        walls[x] = new Wall((float)x * 4, 0, 4);
         walls[x]->setColour({rand() % 255, rand() % 255, rand() % 255});
     }
 //    for (int x = 16; x < 32; x++) {
